Install Spring Boot BOM from source when not present in Maven Central
When updating the boosters as part of the community release process,
we usually don't wait for the BOM to propagate to Maven Central.
That can cause the booster tests to fail for no fault of their own.
In such cases we fall back to downloading the BOM from Github
and installing it locally

Implements: SB-739